+++ app/src/main/java/com/example/per6/hydrationapp/BackendlessSettings.java
@@ -1,0 +1,7 @@
+package com.example.per6.hydrationapp;
+
+class BackendlessSettings {
+
+    public static final String APP_ID = "9BD220AC-2E6E-B5AA-FFA7-5B857386D600";
+    public static final String API_KEY = "52D31C0C-2F4A-A407-FF84-05DE0ED41D00";
+}
