+++ app/src/main/java/com/example/per6/hydrationapp/BluetoothActivity.java
@@ -1,20 +1,28 @@
 package com.example.per6.hydrationapp;
 
 
+import android.Manifest;
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
+import android.annotation.TargetApi;
+import android.app.Activity;
 import android.app.AlertDialog;
 import android.arch.lifecycle.Observer;
 import android.arch.lifecycle.ViewModelProviders;
+import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.le.ScanCallback;
 import android.content.ClipData;
 import android.content.ClipboardManager;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
+import android.provider.Settings;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.StringRes;
@@ -25,12 +33,14 @@ import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.widget.SwipeRefreshLayout;
+import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.PopupMenu;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.SimpleItemAnimator;
+import android.support.v7.widget.Toolbar;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
@@ -44,12 +54,7 @@ import android.view.animation.Animation;
 import android.view.animation.LinearInterpolator;
 import android.view.animation.RotateAnimation;
 import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.SeekBar;
-import android.widget.TextView;
+
 
 import com.example.per6.hydrationapp.ble.BleManager;
 import com.example.per6.hydrationapp.ble.BlePeripheral;
@@ -65,48 +70,48 @@ import no.nordicsemi.android.support.v18.scanner.ScanRecord;
 import static android.content.Context.CLIPBOARD_SERVICE;
 import static android.content.Context.MODE_PRIVATE;
 
-public class BluetoothActivity extends AppCompatActivity implements ScannerStatusFragmentDialog.onScannerStatusCancelListener, BluetoothActivity.ScannerFragmentListener {
+public class BluetoothActivity extends AppCompatActivity implements ScannerStatusFragmentDialog.onScannerStatusCancelListener {
     // Constants
     private final static String TAG = ScannerFragment.class.getSimpleName();
+
+
+    private boolean hasUserAlreadyBeenAskedAboutBluetoothStatus = false;
+
 
     private final static String kPreferences = "Scanner";
     private final static String kPreferences_filtersPanelOpen = "filtersPanelOpen";
 
+    private final static int PERMISSION_REQUEST_COARSE_LOCATION = 1;
+    private static final int kActivityRequestCode_EnableBluetooth = 1;
+
+
     // Data -  Scanned Devices
     private Context context;
-    private ScannerFragmentListener mListener;
     private ScannerViewModel mScannerViewModel;
     private BlePeripheralsAdapter mBlePeripheralsAdapter;
     private SwipeRefreshLayout mSwipeRefreshLayout;
     private Button skip;
+    private FragmentManager fm;
+    private AlertDialog mRequestLocationDialog;
 
     // Data - Dialogs
     private ScannerStatusFragmentDialog mConnectingDialog;
 
     // region Fragment lifecycle
-    public static ScannerFragment newInstance() {
-        return new ScannerFragment();
-    }
-
-    public ScannerFragment() {
-        // Required empty public constructor
-    }
-
-
-
 
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_bluetooth);
-        context=this;
 
-        try {
-            mListener = (ScannerFragmentListener) context;
-        } catch (ClassCastException e) {
-            throw new ClassCastException(context.toString() + " must implement ScannerFragmentListener");
-        }
+        context = this;
+
+
+        fm = getSupportFragmentManager();
+        onActivityCreatedStuff();
+        onViewCreatedStuff();
+
 
         //setHasOptionsMenu(true);
 
@@ -114,17 +119,54 @@ public class BluetoothActivity extends AppCompatActivity implements ScannerStatu
         //setRetainInstance(true);
     }
 
+    private void onActivityCreatedStuff() {
+        mScannerViewModel = ViewModelProviders.of(this).get(ScannerViewModel.class);
+
+        // Scan results
+        mScannerViewModel.getFilteredBlePeripherals().observe(this, new Observer<List<BlePeripheral>>() {
+            @Override
+            public void onChanged(@Nullable List<BlePeripheral> blePeripherals) {
+                mBlePeripheralsAdapter.setBlePeripherals(blePeripherals);
+            }
+        });
+
+        // Scanning
+        mScannerViewModel.getScanningErrorCode().observe(this, new Observer<Integer>() {
+            @Override
+            public void onChanged(@Nullable Integer errorCode) {
+                Log.d(TAG, "Scanning error: " + errorCode);
+
+                if (errorCode != null && errorCode == ScanCallback.SCAN_FAILED_APPLICATION_REGISTRATION_FAILED) {       // Check for known errors
+                    AlertDialog.Builder builder = new AlertDialog.Builder(context);
+                    AlertDialog dialog = builder.setTitle(R.string.dialog_error).setMessage(R.string.bluetooth_scanner_errorregisteringapp)
+                            .setPositiveButton(android.R.string.ok, null)
+                            .show();
+                    DialogUtils.keepDialogOnOrientationChanges(dialog);
+                } else {        // Ask for location permission
+                    requestCoarseLocationPermissionIfNeeded();
+                }
+            }
+        });
+
+        mScannerViewModel.getBlePeripheralsConnectionChanged().observe(this, new Observer<BlePeripheral>() {
+            @Override
+            public void onChanged(@Nullable BlePeripheral blePeripheral) {
+                mBlePeripheralsAdapter.notifyDataSetChanged();
+                if (blePeripheral != null) {
+                    showConnectionStateDialog(blePeripheral);
+                }
+            }
+        });
+        mScannerViewModel.getBlePeripheralDiscoveredServices().observe(this, this::showServiceDiscoveredStateDialog);
+
+        mScannerViewModel.getConnectionErrorMessage().observe(this, this::showConnectionStateError);
+    }
 
 
-    @Override
-    public void onActivityCreated(@NonNull final View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
-
-
+    private void onViewCreatedStuff() {
         if (context != null) {
             // Peripherals recycler view
-            RecyclerView peripheralsRecyclerView = view.findViewById(R.id.peripheralsRecyclerView);
+            RecyclerView peripheralsRecyclerView = findViewById(R.id.peripheralsRecyclerView);
             DividerItemDecoration itemDecoration = new DividerItemDecoration(context, DividerItemDecoration.VERTICAL);
             Drawable lineSeparatorDrawable = ContextCompat.getDrawable(context, R.drawable.simpledivideritemdecoration);
             assert lineSeparatorDrawable != null;
@@ -168,14 +210,14 @@ public class BluetoothActivity extends AppCompatActivity implements ScannerStatu
             peripheralsRecyclerView.setAdapter(mBlePeripheralsAdapter);
 
             // Swipe to refreshAll
-            mSwipeRefreshLayout = view.findViewById(R.id.swipeRefreshLayout);
+            mSwipeRefreshLayout = findViewById(R.id.swipeRefreshLayout);
             mSwipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
                 @Override
                 public void onRefresh() {
                     if (BleManager.getInstance().isAdapterEnabled()) {
                         mScannerViewModel.refresh();
                     } else {
-                        mListener.bluetoothAdapterIsDisabled();
+                        checkPermissions();
                     }
                     new Handler().postDelayed(new Runnable() {
                         @Override
@@ -186,80 +228,28 @@ public class BluetoothActivity extends AppCompatActivity implements ScannerStatu
                 }
             });
 
-            skip = view.findViewById(R.id.skipButton);
-            FragmentManager fragmentManager = getFragmentManager();
+            skip = findViewById(R.id.skipButton);
             skip.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
                     //todo add alert dialog are you sure?
-                    HomepageFragment fragment =  new HomepageFragment();
-                    if (fragmentManager != null) {
-                        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction()
-                                .setCustomAnimations(R.anim.slide_in_left, R.anim.slide_out_right, R.anim.slide_in_right, R.anim.slide_out_left)
-                                .replace(R.id.contentLayout, fragment, "Modules");
-                        fragmentTransaction.addToBackStack(null);
-                        fragmentTransaction.commit();
-                    }
+                    Intent i = new Intent(BluetoothActivity.this, MainActivity.class);
+                    startActivity(i);
+
+
                 }
             });
         }
-
-
     }
 
-    @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
 
-        // ViewModel
-        //FragmentActivity activity = getActivity();
-        mScannerViewModel = ViewModelProviders.of(this).get(ScannerViewModel.class);
-
-        // Scan results
-        mScannerViewModel.getFilteredBlePeripherals().observe(this, new Observer<List<BlePeripheral>>() {
-            @Override
-            public void onChanged(@Nullable List<BlePeripheral> blePeripherals) {
-                mBlePeripheralsAdapter.setBlePeripherals(blePeripherals);
-            }
-        });
-
-        // Scanning
-        mScannerViewModel.getScanningErrorCode().observe(this, new Observer<Integer>() {
-            @Override
-            public void onChanged(@Nullable Integer errorCode) {
-                Log.d(TAG, "Scanning error: " + errorCode);
-
-                if (errorCode != null && errorCode == ScanCallback.SCAN_FAILED_APPLICATION_REGISTRATION_FAILED) {       // Check for known errors
-                    AlertDialog.Builder builder = new AlertDialog.Builder(ScannerFragment.this.getContext());
-                    AlertDialog dialog = builder.setTitle(R.string.dialog_error).setMessage(R.string.bluetooth_scanner_errorregisteringapp)
-                            .setPositiveButton(android.R.string.ok, null)
-                            .show();
-                    DialogUtils.keepDialogOnOrientationChanges(dialog);
-                } else {        // Ask for location permission
-                    mListener.scannerRequestLocationPermissionIfNeeded();
-                }
-            }
-        });
-
-        mScannerViewModel.getBlePeripheralsConnectionChanged().observe(this, new Observer<BlePeripheral>() {
-            @Override
-            public void onChanged(@Nullable BlePeripheral blePeripheral) {
-                mBlePeripheralsAdapter.notifyDataSetChanged();
-                if (blePeripheral != null) {
-                    ScannerFragment.this.showConnectionStateDialog(blePeripheral);
-                }
-            }
-        });
-        mScannerViewModel.getBlePeripheralDiscoveredServices().observe(this, this::showServiceDiscoveredStateDialog);
-
-        mScannerViewModel.getConnectionErrorMessage().observe(this, this::showConnectionStateError);
-    }
 
     @Override
     public void onResume() {
         super.onResume();
+        checkPermissions();
 
-        FragmentActivity activity = getActivity();
+        FragmentActivity activity = this;
         if (activity != null) {
             activity.getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_PAN);
             startScanning();
@@ -271,6 +261,10 @@ public class BluetoothActivity extends AppCompatActivity implements ScannerStatu
         super.onPause();
 
         mScannerViewModel.stop();
+        if (mRequestLocationDialog != null) {
+            mRequestLocationDialog.cancel();
+            mRequestLocationDialog = null;
+        }
     }
 
     @Override
@@ -331,35 +325,30 @@ public class BluetoothActivity extends AppCompatActivity implements ScannerStatu
         }
     }
 
-    private void showConnectionStateDialog(@StringRes int messageId, final BlePeripheral blePeripheral) {
+    private void showConnectionStateDialog(@StringRes int messageId,
+                                           final BlePeripheral blePeripheral) {
         // Show dialog
         final String message = getString(messageId);
         if (mConnectingDialog == null || !mConnectingDialog.isInitialized()) {
             removeConnectionStateDialog();
+            mConnectingDialog = ScannerStatusFragmentDialog.newInstance(message, blePeripheral.getIdentifier());
+            mConnectingDialog.show(fm, "ConnectingDialog");
 
-            FragmentActivity activity = getActivity();
-            if (activity != null) {
-                FragmentManager fragmentManager = getFragmentManager();
-                if (fragmentManager != null) {
-                    mConnectingDialog = ScannerStatusFragmentDialog.newInstance(message, blePeripheral.getIdentifier());
-                    mConnectingDialog.setTargetFragment(this, 0);
-                    mConnectingDialog.show(fragmentManager, "ConnectingDialog");
-                }
-            }
         } else {
             mConnectingDialog.setMessage(message);
         }
     }
 
+
     private void showServiceDiscoveredStateDialog(BlePeripheral blePeripheral) {
-        Context context = this;
+
 
         if (blePeripheral != null && context != null) {
 
             if (blePeripheral.isDisconnected()) {
                 Log.d(TAG, "Abort connection sequence. Peripheral disconnected");
             } else {
-                mListener.startPeripheralModules(blePeripheral.getIdentifier());
+                startPeripheralModules(blePeripheral.getIdentifier());
             }
         }
     }
@@ -367,38 +356,150 @@ public class BluetoothActivity extends AppCompatActivity implements ScannerStatu
     private void showConnectionStateError(String message) {
         removeConnectionStateDialog();
 
-        FragmentActivity activity = getActivity();
-        if (activity != null) {
-            View view = activity.findViewById(android.R.id.content);
-            Snackbar snackbar = Snackbar.make(view, message, Snackbar.LENGTH_LONG);
-            StyledSnackBar.styleSnackBar(snackbar, activity);
-            snackbar.show();
+
+        View view = findViewById(android.R.id.content);
+        Snackbar snackbar = Snackbar.make(view, message, Snackbar.LENGTH_LONG);
+        StyledSnackBar.styleSnackBar(snackbar, context);
+        snackbar.show();
+
+    }
+
+
+    @TargetApi(Build.VERSION_CODES.M)
+    private boolean requestCoarseLocationPermissionIfNeeded() {
+        boolean permissionGranted = true;
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            // Android Marshmallow Permission check 
+            if (checkSelfPermission(Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+                permissionGranted = false;
+                final AlertDialog.Builder builder = new AlertDialog.Builder(this);
+                mRequestLocationDialog = builder.setTitle(R.string.bluetooth_locationpermission_title)
+                        .setMessage(R.string.bluetooth_locationpermission_text)
+                        .setPositiveButton(android.R.string.ok, null)
+                        .setOnDismissListener(dialog -> requestPermissions(new String[]{Manifest.permission.ACCESS_COARSE_LOCATION}, PERMISSION_REQUEST_COARSE_LOCATION))
+                        .show();
+            }
+        }
+        return permissionGranted;
+    }
+
+    private void checkPermissions() {
+
+        final boolean areLocationServicesReadyForScanning = manageLocationServiceAvailabilityForScanning();
+        if (!areLocationServicesReadyForScanning) {
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            mRequestLocationDialog = builder.setMessage(R.string.bluetooth_locationpermission_disabled_text)
+                    .setPositiveButton(android.R.string.ok, null)
+                    .show();
+            //DialogUtils.keepDialogOnOrientationChanges(mRequestLocationDialog);
+        } else {
+            if (mRequestLocationDialog != null) {
+                mRequestLocationDialog.cancel();
+                mRequestLocationDialog = null;
+            }
+
+            // Bluetooth state
+            if (!hasUserAlreadyBeenAskedAboutBluetoothStatus) {     // Don't repeat the check if the user was already informed to avoid showing the "Enable Bluetooth" system prompt several times
+                final boolean isBluetoothEnabled = manageBluetoothAvailability();
+
+                if (isBluetoothEnabled) {
+                    // Request Bluetooth scanning permissions
+                    final boolean isLocationPermissionGranted = requestCoarseLocationPermissionIfNeeded();
+
+                    if (isLocationPermissionGranted) {
+                        // All good. Start Scanning
+                        BleManager.getInstance().start(BluetoothActivity.this);
+                        // Bluetooth was enabled, resume scanning
+                        //mMainFragment.startScanning();
+                    }
+                }
+            }
         }
     }
 
-    @Override
-    public void bluetoothAdapterIsDisabled() {
+    private boolean manageLocationServiceAvailabilityForScanning() {
 
+        boolean areLocationServiceReady = true;
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {        // Location services are only needed to be enabled from Android 6.0
+            int locationMode = Settings.Secure.LOCATION_MODE_OFF;
+            try {
+                locationMode = Settings.Secure.getInt(getContentResolver(), Settings.Secure.LOCATION_MODE);
+
+            } catch (Settings.SettingNotFoundException e) {
+                e.printStackTrace();
+            }
+            areLocationServiceReady = locationMode != Settings.Secure.LOCATION_MODE_OFF;
+        }
+
+        return areLocationServiceReady;
+    }
+
+    private boolean manageBluetoothAvailability() {
+        boolean isEnabled = true;
+
+        // Check Bluetooth HW status
+        int errorMessageId = 0;
+        final int bleStatus = BleUtils.getBleStatus(getBaseContext());
+        switch (bleStatus) {
+            case BleUtils.STATUS_BLE_NOT_AVAILABLE:
+                errorMessageId = R.string.bluetooth_unsupported;
+                isEnabled = false;
+                break;
+            case BleUtils.STATUS_BLUETOOTH_NOT_AVAILABLE: {
+                errorMessageId = R.string.bluetooth_poweredoff;
+                isEnabled = false;      // it was already off
+                break;
+            }
+            case BleUtils.STATUS_BLUETOOTH_DISABLED: {
+                isEnabled = false;      // it was already off
+                // if no enabled, launch settings dialog to enable it (user should always be prompted before automatically enabling bluetooth)
+                Intent enableBluetoothIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
+                startActivityForResult(enableBluetoothIntent, kActivityRequestCode_EnableBluetooth);
+                // execution will continue at onActivityResult()
+                break;
+            }
+        }
+
+        if (errorMessageId != 0) {
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            AlertDialog dialog = builder.setMessage(errorMessageId)
+                    .setPositiveButton(android.R.string.ok, null)
+                    .show();
+            DialogUtils.keepDialogOnOrientationChanges(dialog);
+        }
+
+        return isEnabled;
     }
 
     @Override
-    public void scannerRequestLocationPermissionIfNeeded() {
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
 
+        if (requestCode == kActivityRequestCode_EnableBluetooth) {
+            if (resultCode == Activity.RESULT_OK) {
+                //checkPermissions();
+            } else if (resultCode == Activity.RESULT_CANCELED) {
+                if (!isFinishing()) {
+                    hasUserAlreadyBeenAskedAboutBluetoothStatus = true;     // Remember that
+                    AlertDialog.Builder builder = new AlertDialog.Builder(this);
+                    AlertDialog dialog = builder.setMessage(R.string.bluetooth_poweredoff)
+                            .setPositiveButton(android.R.string.ok, null)
+                            .show();
+                    DialogUtils.keepDialogOnOrientationChanges(dialog);
+                }
+            }
+        }
     }
 
-    @Override
-    public void startPeripheralModules(String singlePeripheralIdentifier) {
+    public void startPeripheralModules(String peripheralIdentifier) {
+        //todo make fragement to which you want app to go
 
+        Intent returnIntent = new Intent();
+        returnIntent.putExtra("peripheralIdentifier", peripheralIdentifier);
+        setResult(Activity.RESULT_OK, returnIntent);
+        finish();
     }
 
-    // region Listeners
-    interface ScannerFragmentListener {
-        void bluetoothAdapterIsDisabled();
-
-        void scannerRequestLocationPermissionIfNeeded();
-
-        void startPeripheralModules(String singlePeripheralIdentifier);
-    }
-
-    // endregion
 }
+
